Server class complete, Add server functionality working.
The server class has been completed as much as is needed for this stage, i.e. no lose ends to tie up there. ServerList is also complete to this point. Add Server functionality is working as far as I can determine but I have not started on Resume, Modify or Delete.